deploy_app/shoutai/requirements.txt
Flask>=2.2.2
mysql-connector-python>=8.0.3.1
python-dotenv>=0.21.0
